requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansible==6.7.0
ansible-core==2.13.13
appnope==0.1.3
APScheduler==3.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
bitarray==2.7.6
blinker==1.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
comm==0.1.3
cryptography==39.0.2
cytoolz==0.12.1
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
executing==1.2.0
Flask==3.0.2
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
importlib-metadata==7.0.0
ipykernel==6.23.0
ipython==8.13.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
lru-dict==1.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
packaging==23.1
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prettytable==3.9.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.6.3
requests==2.28.2
resolvelib==0.8.1
rlp==3.0.0
schedule==1.2.1
six==1.16.0
slack-webhook==1.0.7
stack-data==0.6.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzlocal==5.2
urllib3==1.26.15
wcwidth==0.2.6
web3==6.5.0
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
